review comments
remove redundant .gitignore
push spaceport module into its own source file
